+++ src/pynwb/testing/mock/ecephys.py
@@ -9,6 +9,7 @@ from ...file import ElectrodeTable, NWBFile
 from ...ecephys import ElectricalSeries, ElectrodeGroup, SpikeEventSeries
 from .device import mock_Device
 from .utils import name_generator
+from ...misc import Units
 
 
 def mock_ElectrodeGroup(
@@ -119,3 +120,32 @@ def mock_SpikeEventSeries(
         nwbfile.add_acquisition(spike_event_series)
 
     return spike_event_series
+
+
+def mock_Units(
+    num_units: int = 10,
+    max_spikes_per_unit: int = 10,
+    seed: int = 0,
+    nwbfile: Optional[NWBFile] = None,
+) -> Units:
+
+    units_table = Units()
+    units_table.add_column(name="unit_name", description="a readable identifier for the unit")
+
+    rng = np.random.default_rng(seed=seed)
+
+    times = rng.random(size=(num_units, max_spikes_per_unit)).cumsum(axis=1)
+    spikes_per_unit = rng.integers(1, max_spikes_per_unit, size=num_units)
+
+    spike_times = []
+    for unit_index in range(num_units):
+
+        # Not all units have the same number of spikes
+        spike_times = times[unit_index, : spikes_per_unit[unit_index]]
+        unit_name = f"unit_{unit_index}"
+        units_table.add_unit(spike_times=spike_times, unit_name=unit_name)
+
+    if nwbfile is not None:
+        nwbfile.units = units_table
+
+    return units_table